requirements.txt
matplotlib
scikit-learn
fire
torchinfo
numpy
plotext
pandas
wandb
lightning
datasets
huggingface_hub